refactor-compact-tiles * Move TileRow directory out of components
