+++ src/dashboards/Compact/TileRow/PlatformInfo/PlatformInfo.module.scss
