+++ src/dashboards/Compact/TileRow/PlatformInfo/PlatformInfo.tsx
