+++ src/dashboards/Compact/TileRow/TileRow.module.scss
@@ -1,4 +1,4 @@
-@use '../../../../assets/styles/Variables.scss';
+@use 'src/assets/styles/Variables';
 
 .Tilerow {
     display: flex;
