+++ src/dashboards/Compact/TileRow/TileRow.tsx
@@ -2,8 +2,8 @@ import React, { Fragment } from 'react'
 import { format, isSameDay, isToday, formatISO } from 'date-fns'
 import { nb } from 'date-fns/locale'
 import { Heading3 } from '@entur/typography'
-import { TileSubLabel } from '../../../../types'
-import { SubLabelIcon } from '../../../../components/SubLabelIcon/SubLabelIcon'
+import { TileSubLabel } from '../../../types'
+import { SubLabelIcon } from '../../../components/SubLabelIcon/SubLabelIcon'
 import { PlatformInfo } from './PlatformInfo/PlatformInfo'
 import classes from './TileRow.module.scss'
